Update to Camera
Added a camera that follows player. Not sure it's the best solution or if it can be tweaked to allow it to not be completely locked to player. Will look to see if it can have a small dead zone.
Also added a sprite to player to make it easier to tell which is Player
Extended the base platform to enable player more room to move.