Added the enemy, health, and Game Over screen
Added the enemy that just moves back and forth for now. Whenever the player hits the enemy, the health in the top left will decrease. If the health reaches 0 then the game over screen will appear, stopping everything from moving. If the retry button is clicked then it completely resets everything and lets you try again.