Added a couple of MP3 files.
added a couple of mp3 files to the assests.
I have not connected them to the game yet.
I don't know if the theme works well for the feel of the game... maybe it would be better as an intro or before game play tune.
the dead sound was something that I thought could work well with an end of game screen.